*Check diferent scenarions and fix a bugs
*Added tokenAuth middlewar
*Added utilits
*Added ban to deleting accout when:
-driver has assigned load
-shipper has assigned load
*Baned  delete\update trucks, when one of them in status "OL" and not
asigned
*Baned delete\update loads, when status of it not 'NEW'
*Baned delet  loads, when status of it not "NEW" or "SHIPPED"
*Banned assign 2 trucks at the same time
*Baned assign another truck, when some truck in statul "OL"
*Added driver functionality to interact with load assigned by him
*Banned oportunity to add truck, if some one in status "OL"
*Added handlign of errors
*Added logs
*Added auto logs on load, when smth i going on with that

+++ homework3/semi-uber/Server/api/handlers/userHandlers/userDeteleHandler.js
@@ -2,25 +2,39 @@ const log4js = require("log4js");
 const logger = log4js.getLogger();
 const { User } = require("../../Schemas/user.schema");
 const { Load } = require("../../Schemas/load.schema");
+const { Truck } = require("../../Schemas/truck.schema");
 const errorHandler = require("../error.handler");
+const checkForEdit = require("../../utilits/checkforEdit");
 
-function userDeleteHandler(req, res) {
+async function userDeleteHandler(req, res) {
   const { id: userID } = req.params;
 
-  User.findByIdAndDelete(userID)
-    .then(() => Load.remove({ created_by: userID }))
-    .then(() => {
-      logger.info(
-        "User's account and all custom data was successfully deleted"
-      );
-      res.json({
-        success: true,
-        data: {},
-        message: "User's account all custom data was successfully deleted"
-      });
-    })
-    .catch(error => {
-      return errorHandler(error.message, res);
-    });
+  const editCheck = await checkForEdit(userID);
+  if (editCheck === null) {
+    return errorHandler("Can not delete account!", res);
+  }
+  const user = await User.findByIdAndDelete(userID);
+
+  if (user.status === "Shipper") {
+    try {
+      await Load.remove({ created_by: userID });
+    } catch (error) {
+      return errorHandler("Can not delete user data");
+    }
+  } else {
+    try {
+      Truck.remove({ created_by: userID });
+    } catch (error) {
+      return errorHandler("Can not delete user data");
+    }
+  }
+  logger.info("User data was deleted!");
+  res.json({
+    success: true,
+    data: {
+      message: "User's account all custom data was successfully deleted"
+    },
+    error: null
+  });
 }
 module.exports = userDeleteHandler;
